chore(packages): restructure; chore(core,pkgs): add makefile to test packages and core.
