+++ server/core/diagram/c4container/compression/blocksplitter.go
