+++ server/core/diagram/c4container/compression/cache.go
