+++ server/core/diagram/c4container/compression/deflate.go
