+++ server/core/diagram/c4container/compression/gzip_container.go
