+++ server/core/diagram/c4container/compression/hash.go
