+++ server/core/diagram/c4container/compression/katajainen.go
