+++ server/core/diagram/c4container/compression/lz77.go
