+++ server/core/diagram/c4container/compression/squeeze.go
