+++ server/core/diagram/c4container/compression/tree.go
