+++ server/core/diagram/c4container/compression/util.go
