+++ server/core/diagram/c4container/compression/zlib_container.go
