--- server/core/domain/c4container/utils.go
@@ -1,9 +1,0 @@
-package c4container
-
-import "strings"
-
-func stringCleaner(s string) string {
-	s = strings.TrimSpace(s)
-	s = strings.ReplaceAll(s, "\n", "\\n")
-	return s
-}
