+++ server/core/internal/utils/svg_test.go
