+++ server/core/pkg/gcpsecretsmanager/go.mod
@@ -5,7 +5,6 @@ go 1.19
 require (
 	cloud.google.com/go/secretmanager v1.10.0
 	github.com/googleapis/gax-go/v2 v2.7.0
-	google.golang.org/api v0.103.0
 )
 
 require (
@@ -21,6 +20,7 @@ require (
 	golang.org/x/oauth2 v0.0.0-20221014153046-6fdb5e3db783 // indirect
 	golang.org/x/sys v0.0.0-20220728004956-3c1f35247d10 // indirect
 	golang.org/x/text v0.4.0 // indirect
+	google.golang.org/api v0.103.0 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/genproto v0.0.0-20221201164419-0e50fba7f41c // indirect
 	google.golang.org/grpc v1.51.0 // indirect
