+++ server/core/pkg/httpclient/go.mod
@@ -1,0 +1,3 @@
+module github.com/kislerdm/diagramastext/server/core/pkg/httpclient
+
+go 1.19
