+++ server/core/pkg/openai/go.mod
@@ -1,0 +1,3 @@
+module github.com/kislerdm/diagramastext/server/core/pkg/openai
+
+go 1.19
