--- server/core/port/diagramhandler.go
@@ -1,8 +1,0 @@
-package port
-
-import (
-	"context"
-)
-
-// DiagramHandler handler to generate a diagram given the input.
-type DiagramHandler func(ctx context.Context, input Input) (Output, error)
